2022 MLB wOBA Data

Weighted on-base average (wOBA) for hitters from the 2022 Major League Baseball (MLB) season. Standard errors are calculated by modeling hitting outcomes as multinomially distributed and plugging in empirical proportions as the “true” outcome probabilities. To handle small sample sizes, standard errors are lower bounded by the errors that would be obtained using league-wide proportions rather than the plug-in estimates.

Format

wOBA is a data frame with 688 rows and 6 columns:

FanGraphsID

The hitter's FanGraphs identifier.

Name

The hitter's name.

Team

The hitter's MLB team (given as a three-letter code) or NA if the hitter played for multiple teams.

PA

The hitter's number of plate appearances over the 2022 season.

x

The hitter's wOBA over the 2022 season.

s

The standard error for the hitter's wOBA.

Source

<https://fangraphs.com>

Examples

data(wOBA)
summary(wOBA)
#>  FanGraphsID            Name                Team           PA       
#>  Length:688         Length:688         OAK    : 27   Min.   :  1.0  
#>  Class :character   Class :character   LAA    : 26   1st Qu.: 65.0  
#>  Mode  :character   Mode  :character   PIT    : 25   Median :220.5  
#>                                        CIN    : 24   Mean   :264.6  
#>                                        SFG    : 24   3rd Qu.:443.0  
#>                                        (Other):483   Max.   :724.0  
#>                                        NA's   : 79                  
#>        x                s          
#>  Min.   :0.0000   Min.   :0.02000  
#>  1st Qu.:0.2487   1st Qu.:0.02500  
#>  Median :0.2890   Median :0.03500  
#>  Mean   :0.2790   Mean   :0.06052  
#>  3rd Qu.:0.3230   3rd Qu.:0.06325  
#>  Max.   :1.0360   Max.   :0.73300  
#>